% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGEN.R
\name{dataGEN}
\alias{dataGEN}
\title{Function to prepare data from other cubes and accessible datasets}
\usage{
dataGEN(arg1, ...)
}
\arguments{
\item{arg1}{Explain parameters here}

\item{...}{Explain parameters here}
}
\value{
Explain what outputs are returned
}
\description{
All functions developed by Stellenbosch BioMath for the b-cubed project
}
\examples{
dataGEN(1)
dataGEN('Also works for character strings')
}
