% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissim.R
\name{dissim}
\alias{dissim}
\title{Function to generate dissimilarity cube}
\usage{
dissim(arg1, ...)
}
\arguments{
\item{arg1}{Explain parameters here}

\item{...}{Explain parameters here}
}
\value{
Explain what outputs are returned
}
\description{
All functions developed by Stellenbosch BioMath for the b-cubed project
}
\examples{
dissim(1)
dissim('Also works for character strings')
}
