% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invasib.R
\name{invasib}
\alias{invasib}
\title{Function to generate invasibility cube}
\usage{
invasib(arg1, ...)
}
\arguments{
\item{arg1}{Explain parameters here}

\item{...}{Explain parameters here}
}
\value{
Explain what outputs are returned
}
\description{
All functions developed by Stellenbosch BioMath for the b-cubed project
}
\examples{
invasib(1)
invasib('Also works for character strings')
}
