% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Function to simulate ecological data}
\usage{
simData(
  startDate = "2020-01-01",
  endDate = "2020-12-31",
  yMin = -25.7,
  yMax = -22.3,
  xMin = 30.7,
  xMax = 32.2,
  speciesList = paste0("species_", 1:10),
  classList = c("alien", "native"),
  envVars = 4,
  spatialRes = 10,
  returnOption = 1,
  plot = FALSE,
  globalSave = TRUE
)
}
\arguments{
\item{startDate}{The starting date for the simulated data in "YYYY-MM-DD" format (default "2020-01-01").}

\item{endDate}{The ending date for the simulated data in "YYYY-MM-DD" format (default "2020-12-31").}

\item{yMin}{The minimum latitude (southern limit) of the spatial extent (default -25.7).}

\item{yMax}{The maximum latitude (northern limit) of the spatial extent (default -22.3).}

\item{xMin}{The minimum longitude (western limit) of the spatial extent (default 30.7).}

\item{xMax}{The maximum longitude (eastern limit) of the spatial extent (default 32.2).}

\item{speciesList}{A vector of species names to be included in the simulation (default c("species_1" to "species_10")).}

\item{classList}{A vector indicating the classification of each species as either "alien" or "native" (default c("alien", "native")).}

\item{envVars}{The number of environmental variables to simulate (default 4).}

\item{spatialRes}{The spatial resolution in kilometers (default 10 km).}

\item{returnOption}{An integer that determines the type of output:
\itemize{
\item 1 for a summary table of species by site,
\item 2 for a SpatRaster object with the full dataset,
\item 3 for a distance matrix of sites.
Other options can be added as needed.
}}

\item{plot}{A logical indicating whether to plot the results; \code{TRUE} generates a plot of species counts across the spatial grid (default FALSE).}

\item{globalSave}{A logical indicating whether to save the grid data to the global environment; \code{TRUE} saves the grid to .GlobalEnv (default FALSE).}
}
\value{
Depending on the \code{returnOption} parameter, the function can return:
\itemize{
\item A summary table (data.frame) of species occurrences by site.
\item A \code{SpatRaster} object representing the full dataset.
\item A distance matrix of sites.
If \code{plot} is TRUE, a plot will be generated for visualizing the distribution of species counts.
}
}
\description{
This function creates a simulated dataset representing ecological observations
over a specified spatial and temporal range. It allows for the generation of species
occurrence data along with environmental variables across a defined spatial grid.
Optionally, the grid can be saved to the global environment.
}
\examples{
set.seed(123)  # For reproducibility
simData_df = simData(returnOption = 1, plot = TRUE)
head(simData_df)
# To check the grid data saved to global environment
head(simData_grid)
}
